requirements.txt
kivy==2.3.0
pillow
reportlab
androidstorage4kivy>=0.1.0
plyer
